Initial comit
A 2D truss solver library. Elements of the truss are bar elements.
Each bar element consists of 2 nodes. With each node having 2
degrees of freedom.

The Truss class is responsible for solving the model. The displacements
of each node are caclulated along with the reactions at the supported
nodes. The stress of each element is also calculated. The driver code
is located at main.py.

Except the solver the repository also includes a visualization package.
The undeformed truss is plotted along with the nodes and the forces
applied at each node. If a solution of the truss model is available, the
deformed truss is plotted as well using a scaling factor for visualization
purposes.

Signed-off-by: Giorgos Tharropoulos <tharropoulos@gmail.com>